+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to this project will be documented in this file.
 - Fixed Egg Carrier starting transformed when spawned inside.
 - The game will now grant the egg checks even if your garden is full.
 - Fixed Final Egg tower entrance sometimes teleporting the character to the lower floor.
+- Fixed Final Egg entrance sometimes changing character to Sonic.
 
 ### Logic